requirements.txt
fsspec
argparse
python-dotenv
